modify waitinterupt parameter
change the second parameter in this line "wait_event_interruptible(data->wait, (data->buffer_status & SENSOR_MASK));  // turn irq"
before I change it, when I run the cat or the test program, only display the first value in the magnetic. and when calling the read file function for second time. it will be stuck. it's wired, and when I cat the goldfish_sensor, it's also in the stuck.