fix the bug about the write function
I dont know why this problem happened, but I fix it now.
In the control_activate function, I just moved the open file function before the qemu_channel function which was written in this function at the very beginning.
and when i started up the emulator again, it works. I mean, the app can display the correct data from sensor.